+++ src/store/modules/auth/actions.js
@@ -19,10 +19,28 @@ export function signInSuccess(token, user) {
   };
 }
 
-export function signUpRequest(name, email, password) {
+export function signUpRequest(
+  name,
+  email,
+  password,
+  street,
+  number,
+  city,
+  state,
+  postal_code
+) {
   return {
     type: '@auth/SIGN_UP_REQUEST',
-    payload: { name, email, password },
+    payload: {
+      name,
+      email,
+      password,
+      street,
+      number,
+      city,
+      state,
+      postal_code,
+    },
   };
 }
